Module casbin::error

Structs

AdapterError
AdapterError error represents any type of errors in adapter's execution

Enums

Error
General casbin error

ModelError
ModelError represents any type of errors in model configuration

PolicyError
PolicyError represents any type of errors in policy

RbacError
RBAC error represents any type of errors in RBAC role manager

RequestError
RequestError represents any type of errors in coming request