Function casbin::rhai::config::hashing::get_hashing_seed

pub fn get_hashing_seed() -> &'static Option<[u64; 4]>

Get the current hashing Seed.

If the seed is not yet defined, the RHAI_HASHING_SEED environment variable (if any) is used.

Otherwise, the hashing seed is randomized to protect against DOS attacks.

See rhai::config::hashing::set_hashing_seed for more.