Module casbin::rhai::packages

Module containing all built-in packages available to Rhai, plus facilities to define custom packages.

Macros

def_package

Macro that makes it easy to define a package (which is basically a shared module) and register functions into it.

Structs

ArithmeticPackage

Basic arithmetic package.

BasicArrayPackage

Package of basic array utilities.

BasicBlobPackage

Package of basic BLOB utilities.

BasicFnPackage

Package of basic function pointer utilities.

BasicIteratorPackage

Package of basic range iterators

BasicMapPackage

Package of basic object map utilities.

BasicMathPackage

Basic mathematical package.

BasicStringPackage

Package of basic string utilities (e.g. printing)

BasicTimePackage

Package of basic timing utilities.

BitFieldPackage

Package of basic bit-field utilities.

CorePackage

Core package containing basic facilities.

LanguageCorePackage

Package of core language features.

LogicPackage

Package of basic logic operators.

MoreStringPackage

Package of additional string utilities over BasicStringPackage

StandardPackage

Standard package containing all built-in features.

Traits

Package

Trait that all packages must implement.