Enum casbin::rhai::plugin::CallableFunction

#[non_exhaustive]
pub enum CallableFunction {
    Pure {
        func: Arc<dyn Fn(Option<NativeCallContext<'_>>, &mut [&mut Dynamic]) + Sync + Send + 'static>,
        has_context: bool,
        is_pure: bool,
    },
    Method {
        func: Arc<dyn Fn(Option<NativeCallContext<'_>>, &mut [&mut Dynamic]) + Sync + Send + 'static>,
        has_context: bool,
        is_pure: bool,
    },
    Iterator {
        func: Arc<dyn Fn(Dynamic) + Sync + Send + 'static>,
    },
    Plugin {
        func: Arc<dyn PluginFunction + Sync + Send + 'static>,
    },
}

(internals) A type encapsulating a function callable by Rhai. Exported under the internals feature only.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

Pure

Fields

func: Arc<dyn Fn(Option<NativeCallContext<'_>>, &mut [&mut Dynamic]) + Sync + Send + 'static>

Shared function pointer.

has_context: bool

Does the function take a NativeCallContext parameter?

is_pure: bool

This is a dummy field and is not used.

A pure native Rust function with all arguments passed by value.

Method

Fields

func: Arc<dyn Fn(Option<NativeCallContext<'_>>, &mut [&mut Dynamic]) + Sync + Send + 'static>

Shared function pointer.

has_context: bool

Does the function take a NativeCallContext parameter?

is_pure: bool

Allow operating on constants?

A native Rust object method with the first argument passed by reference, and the rest passed by value.

Iterator

Fields

func: Arc<dyn Fn(Dynamic) + Sync + Send + 'static>

Shared function pointer.

An iterator function.

Plugin

Fields

func: Arc<dyn PluginFunction + Sync + Send + 'static>

Shared function pointer.

A plugin function,

Implementations

impl CallableFunction

pub fn is_pure(&self) -> bool

Is this a pure native Rust function?

pub fn is_method(&self) -> bool

Is this a native Rust method function?

pub const fn is_iter(&self) -> bool

Is this an iterator function?

pub const fn is_script(&self) -> bool

Is this a script-defined function?

pub const fn is_plugin_fn(&self) -> bool

Is this a plugin function?

pub const fn is_native(&self) -> bool

Is this a native Rust function?

pub fn has_context(&self) -> bool

Is there a NativeCallContext parameter?

pub fn access(&self) -> FnAccess

Get the access mode.

pub fn get_native_fn(
&self
) -> Option<&Arc<dyn Fn(Option<NativeCallContext<'_>>, &mut [&mut Dynamic]) + Sync + Send + 'static>>

Get a shared reference to a native Rust function.

pub fn get_encapsulated_environ(&self) -> Option<&EncapsulatedEnviron>

Get a reference to the shared encapsulated environment of the function definition.

Not available under no_function or no_module.

pub fn get_iter_fn(&self) -> Option<&(dyn Fn(Dynamic) + Sync + Send + 'static)>

Get a reference to an iterator function.

pub fn get_plugin_fn(
&self
) -> Option<&Arc<dyn PluginFunction + Sync + Send + 'static>>

Get a shared reference to a plugin function.