casbin::rhai::plugin
Enum RhaiFunc

#[non_exhaustive]
pub enum RhaiFunc {
    Pure {
        func: Arc<dyn Fn(Option<NativeCallContext<'_>>, &mut [&mut Dynamic]) -> Result<Dynamic, Box<EvalAltResult>> + Send + Sync>,
        has_context: bool,
        is_pure: bool,
        is_volatile: bool,
    },
    Method {
        func: Arc<dyn Fn(Option<NativeCallContext<'_>>, &mut [&mut Dynamic]) -> Result<Dynamic, Box<EvalAltResult>> + Send + Sync>,
        has_context: bool,
        is_pure: bool,
        is_volatile: bool,
    },
    Iterator {
        func: Arc<dyn Fn(Dynamic) -> Box<dyn Iterator<Item = Result<Dynamic, Box<EvalAltResult>>>> + Send + Sync>,
    },
    Plugin {
        func: Arc<dyn PluginFunc + Send + Sync>,
    },
}

(internals) A type encapsulating a function callable by Rhai. Exported under the internals feature only.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

Pure

A pure native Rust function with all arguments passed by value.

Fields

func: Arc<dyn Fn(Option<NativeCallContext<'_>>, &mut [&mut Dynamic]) -> Result<Dynamic, Box<EvalAltResult>> + Send + Sync>

Shared function pointer.

has_context: bool

Does the function take a NativeCallContext parameter?

is_pure: bool

This is a dummy field and is not used.

is_volatile: bool

Is this function volatile?

A volatile function does not guarantee the same result for the same input(s).

Method

A native Rust object method with the first argument passed by reference, and the rest passed by value.

Fields

func: Arc<dyn Fn(Option<NativeCallContext<'_>>, &mut [&mut Dynamic]) -> Result<Dynamic, Box<EvalAltResult>> + Send + Sync>

Shared function pointer.

has_context: bool

Does the function take a NativeCallContext parameter?

is_pure: bool

Allow operating on constants?

is_volatile: bool

Is this function volatile?

A volatile function does not guarantee the same result for the same input(s).

Iterator

An iterator function.

Fields

func: Arc<dyn Fn(Dynamic) -> Box<dyn Iterator<Item = Result<Dynamic, Box<EvalAltResult>>>> + Send + Sync>

Shared function pointer.

Plugin

A plugin function,

Fields

func: Arc<dyn PluginFunc + Send + Sync>

Shared function pointer.

Implementations

impl RhaiFunc

pub fn is_pure(&self) -> bool

Is this a pure native Rust function?

pub fn is_method(&self) -> bool

Is this a native Rust method function?

pub const fn is_iter(&self) -> bool

Is this an iterator function?

pub const fn is_script(&self) -> bool

Is this a script-defined function?

pub const fn is_plugin_fn(&self) -> bool

Is this a plugin function?

pub const fn is_native(&self) -> bool

Is this a native Rust function?

pub fn has_context(&self) -> bool

Is there a NativeCallContext parameter?

pub fn is_volatile(&self) -> bool

Is this function volatile?

A volatile function does not guarantee the same result for the same input(s).

pub fn access(&self) -> FnAccess

Get the access mode.

pub fn get_native_fn( &self, ) -> Option<&Arc<dyn Fn(Option<NativeCallContext<'_>>, &mut [&mut Dynamic]) -> Result<Dynamic, Box<EvalAltResult>> + Send + Sync>>

Get a shared reference to a native Rust function.

pub fn get_encapsulated_environ(&self) -> Option<&EncapsulatedEnviron>

Get a reference to the shared encapsulated environment of the function definition.

Not available under no_function or no_module.

pub fn get_iter_fn( &self, ) -> Option<&(dyn Fn(Dynamic) -> Box<dyn Iterator<Item = Result<Dynamic, Box<EvalAltResult>>>> + Send + Sync + 'static)>

Get a reference to an iterator function.

pub fn get_plugin_fn(&self) -> Option<&Arc<dyn PluginFunc + Send + Sync>>

Get a shared reference to a plugin function.