casbin::rhai::plugin
Macro register_exported_fn

register_exported_fn!() { /* proc-macro */ }

Macro to register a plugin function (defined via #[export_fn]) into an Engine.

Deprecated

This macro is deprecated as it performs no additional value.

This method will be removed in the next major version.