Struct casbin::rhai::Expression

pub struct Expression<'a>(_);

An expression sub-tree in an AST.

Implementations

impl Expression<'_>

pub fn eval_with_context( &self, context: &mut EvalContext<'_, '_, '_, '_, '_, '_> ) -> Result<Dynamic, Box<EvalAltResult, Global>>

Evaluate this expression tree within an evaluation context.

WARNING - Low Level API

This function is very low level. It evaluates an expression from an AST.

pub fn eval_with_context_raw( &self, context: &mut EvalContext<'_, '_, '_, '_, '_, '_>, rewind_scope: bool ) -> Result<Dynamic, Box<EvalAltResult, Global>>

๐Deprecated: This API is NOT deprecated, but it is considered volatile and may change in the future.

Evaluate this expression tree within an evaluation context.

The following option is available:

whether to rewind the Scope after evaluation if the expression is a [StmtBlock][crate::ast::StmtBlock]

WARNING - Unstable API

This API is volatile and may change in the future.

WARNING - Low Level API

This function is extremely low level. It evaluates an expression from an AST.

pub fn get_string_value(&self) -> Option<&str>

Get the value of this expression if it is a variable name or a string constant.

Returns None also if the constant is not of the specified type.

pub const fn position(&self) -> Position

Get the position of this expression.

pub fn get_literal_value<T>(&self) -> Option<T>where T: Variant,

Get the value of this expression if it is a literal constant.

Supports INT, [FLOAT][crate::FLOAT], (), char, bool and ImmutableString.

Returns None also if the constant is not of the specified type.

impl Expression<'_>

pub fn get_variable_name(&self) -> Option<&str>

๐Deprecated since 1.4.0: use get_string_value instead

If this expression is a variable name, return it. Otherwise None.

Deprecated

This method is deprecated. Use get_string_value instead.

This method will be removed in the next major version.