Struct casbin::rhai::ImmutableString

pub struct ImmutableString(_);

The system immutable string type.

An ImmutableString wraps an Rc<String> (or Arc<String> under the sync feature) so that it can be simply shared and not cloned.

Examples

use rhai::ImmutableString;

let s1: ImmutableString = "hello".into();

// No actual cloning of the string is involved below.
let s2 = s1.clone();
let s3 = s2.clone();

assert_eq!(s1, s2);

// Clones the underlying string (because it is already shared) and extracts it.
let mut s: String = s1.into_owned();

// Changing the clone has no impact on the previously shared version.
s.push_str(", world!");

// The old version still exists.
assert_eq!(s2, s3);
assert_eq!(s2.as_str(), "hello");

// Not equals!
assert_ne!(s2.as_str(), s.as_str());
assert_eq!(s, "hello, world!");

Implementations

impl ImmutableString

pub fn into_owned(self) -> String

Consume the ImmutableString and convert it into a String. If there are other references to the same string, a cloned copy is returned.

pub fn make_mut(&mut self) -> &mut String

Make sure that the ImmutableString is unique (i.e. no other outstanding references). Then return a mutable reference to the String.