Struct casbin::rhai::Position

pub struct Position { /* private fields */ }

A location (line number + character position) in the input script.

Limitations

In order to keep footprint small, both line number and character position have 16-bit resolution, meaning they go up to a maximum of 65,535 lines and 65,535 characters per line.

Advancing beyond the maximum line length or maximum number of lines is not an error but has no effect.

Implementations

impl Position

pub const fn new_const(line: u16, position: u16) -> Option<Position>

👎Deprecated since 1.6.0: use new instead

Create a new Position.

If line is zero, then [None] is returned.

If position is zero, then it is at the beginning of a line.

Deprecated

This function is deprecated. Use new (which panics when line is zero) instead.

This method will be removed in the next major version.

impl Position

pub const NONE: Position = Self{ line: 0, pos: 0,}

A Position representing no position.

pub const START: Position = Self{ line: 1, pos: 0,}

A Position representing the first position.

pub const fn new(line: u16, position: u16) -> Position

Create a new Position.

line must not be zero.

If position is zero, then it is at the beginning of a line.

Panics

Panics if line is zero.

pub const fn line(self) -> Option<usize>

Get the line number (1-based), or [None] if there is no position.

Always returns [None] under no_position.

pub const fn position(self) -> Option<usize>

Get the character position (1-based), or [None] if at beginning of a line.

Always returns [None] under no_position.

pub const fn is_beginning_of_line(self) -> bool

Is this Position at the beginning of a line?

Always returns false under no_position.

pub const fn is_none(self) -> bool

Is there no Position?

Always returns true under no_position.

pub const fn or_else(self, pos: Position) -> Position

Returns an fallback Position if it is NONE?

Always returns the fallback under no_position.