Trait casbin::rhai::RegisterNativeFunction

pub trait RegisterNativeFunction<A, const N: usize, const X: bool, R, const F: bool>where
    A: 'static,
    R: 'static,{
    fn into_callable_function(
        self,
        name: SmartString<LazyCompact>,
        no_const: bool
    ) -> CallableFunction;
    fn param_types() -> [TypeId; N];

    fn num_params() -> usize { ... }
    fn has_context() -> bool { ... }
}

Trait to register custom Rust functions.

Type Parameters

A - a tuple containing parameter types, with &mut T represented by Mut<T>.

N - a constant generic containing the number of parameters, must be consistent with ARGS.

X - a constant boolean generic indicating whether there is a NativeCallContext parameter.

R - return type of the function; if the function returns Result, it is the unwrapped inner value type.

F - a constant boolean generic indicating whether the function is fallible (i.e. returns Result<T, Box<EvalAltResult>>).

Required Methods

fn into_callable_function(
self,
name: SmartString<LazyCompact>,
no_const: bool
) -> CallableFunction

Convert this function into a CallableFunction.

fn param_types() -> [TypeId; N]

Get the type ID’s of this function’s parameters.

Provided Methods

fn num_params() -> usize

Get the number of parameters for this function.

fn has_context() -> bool

Is there a NativeCallContext parameter for this function?