Enum casbin::rhai::EvalAltResult

#[non_exhaustive]
pub enum EvalAltResult {

    ErrorSystem(String, Box<dyn Error + 'static + Send + Sync, Global>),
    ErrorParsing(ParseErrorType, Position),
    ErrorVariableNotFound(String, Position),
    ErrorFunctionNotFound(String, Position),
    ErrorInFunctionCall(String, String, Box<EvalAltResult, Global>, Position),
    ErrorModuleNotFound(String, Position),
    ErrorInModule(String, Box<EvalAltResult, Global>, Position),
    ErrorUnboundThis(Position),
    ErrorMismatchDataType(String, String, Position),
    ErrorMismatchOutputType(String, String, Position),
    ErrorArrayBounds(usize, i32, Position),
    ErrorStringBounds(usize, i32, Position),
    ErrorBitFieldBounds(usize, i32, Position),
    ErrorIndexingType(String, Position),
    ErrorFor(Position),
    ErrorDataRace(String, Position),
    ErrorAssignmentToConstant(String, Position),
    ErrorDotExpr(String, Position),
    ErrorArithmetic(String, Position),
    ErrorTooManyOperations(Position),
    ErrorTooManyModules(Position),
    ErrorStackOverflow(Position),
    ErrorDataTooLarge(String, Position),
    ErrorTerminated(Dynamic, Position),
    ErrorRuntime(Dynamic, Position),
    LoopBreak(bool, Position),
    Return(Dynamic, Position),

}

Evaluation result.

All wrapped Position values represent the location in the script where the error occurs.

Thread Safety

Currently, EvalAltResult is neither Send nor Sync. Turn on the sync feature to make it Send + Sync.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

ErrorSystem(String, Box<dyn Error + 'static + Send + Sync, Global>)

System error. Wrapped values are the error message and the internal error.

ErrorParsing(ParseErrorType, Position)

Syntax error.

ErrorVariableNotFound(String, Position)

Usage of an unknown variable. Wrapped value is the variable name.

ErrorFunctionNotFound(String, Position)

Call to an unknown function. Wrapped value is the function signature.

ErrorInFunctionCall(String, String, Box<EvalAltResult, Global>, Position)

An error has occurred inside a called function. Wrapped values are the function name, function source, and the interior error.

ErrorModuleNotFound(String, Position)

Usage of an unknown module. Wrapped value is the module name.

ErrorInModule(String, Box<EvalAltResult, Global>, Position)

An error has occurred while loading a module. Wrapped value are the module name and the interior error.

ErrorUnboundThis(Position)

Access to this that is not bound.

ErrorMismatchDataType(String, String, Position)

Data is not of the required type. Wrapped values are the type requested and type of the actual result.

ErrorMismatchOutputType(String, String, Position)

Returned type is not the same as the required output type. Wrapped values are the type requested and type of the actual result.

ErrorArrayBounds(usize, i32, Position)

Array access out-of-bounds. Wrapped values are the current number of elements in the array and the index number.

ErrorStringBounds(usize, i32, Position)

String indexing out-of-bounds. Wrapped values are the current number of characters in the string and the index number.

ErrorBitFieldBounds(usize, i32, Position)

Bit-field indexing out-of-bounds. Wrapped values are the current number of bits in the bit-field and the index number.

ErrorIndexingType(String, Position)

Trying to index into a type that has no indexer function defined. Wrapped value is the type name.

ErrorFor(Position)

The for statement encounters a type that is not an iterator.

ErrorDataRace(String, Position)

Data race detected when accessing a variable. Wrapped value is the variable name.

ErrorAssignmentToConstant(String, Position)

Assignment to a constant variable. Wrapped value is the variable name.

ErrorDotExpr(String, Position)

Inappropriate property access. Wrapped value is the property name.

ErrorArithmetic(String, Position)

Arithmetic error encountered. Wrapped value is the error message.

ErrorTooManyOperations(Position)

Number of operations over maximum limit.

ErrorTooManyModules(Position)

Modules over maximum limit.

ErrorStackOverflow(Position)

Call stack over maximum limit.

ErrorDataTooLarge(String, Position)

Data value over maximum size limit. Wrapped value is the type name.

ErrorTerminated(Dynamic, Position)

The script is prematurely terminated. Wrapped value is the termination token.

ErrorRuntime(Dynamic, Position)

Run-time error encountered. Wrapped value is the error token.

LoopBreak(bool, Position)

Breaking out of loops - not an error if within a loop. The wrapped value, if true, means breaking clean out of the loop (i.e. a break statement). The wrapped value, if false, means breaking the current context (i.e. a continue statement).

Return(Dynamic, Position)

Not an error: Value returned from a script via the return keyword. Wrapped value is the result value.

Implementations

impl EvalAltResult

#[must_use]
pub const fn is_pseudo_error(&self) -> bool

Is this a pseudo error? A pseudo error is one that does not occur naturally.

LoopBreak and Return are pseudo errors.

#[must_use]
pub const fn is_catchable(&self) -> bool

Can this error be caught?

#[must_use]
pub const fn is_system_exception(&self) -> bool

Is this error a system exception?

#[must_use]
pub const fn position(&self) -> Position

Get the position of this error.

pub fn clear_position(&mut self) -> &mut EvalAltResult

Remove the position information from this error.

The position of this error is set to NONE afterwards.

pub fn take_position(&mut self) -> Position

Remove the position information from this error and return it.

The position of this error is set to NONE afterwards.

pub fn set_position(&mut self, new_position: Position) -> &mut EvalAltResult

Override the position of this error.