Enum casbin::rhai::EvalAltResult

#[non_exhaustive]pub enum EvalAltResult {
    ErrorParsing(ParseErrorType, Position),
    ErrorReadingScriptFile(PathBuf, Position, Error),
    ErrorFunctionNotFound(String, Position),
    ErrorInFunctionCall(String, Box<EvalAltResult>, Position),
    ErrorUnboundThis(Position),
    ErrorBooleanArgMismatch(String, Position),
    ErrorCharMismatch(Position),
    ErrorArrayBounds(usize, i32, Position),
    ErrorStringBounds(usize, i32, Position),
    ErrorIndexingType(String, Position),
    ErrorNumericIndexExpr(Position),
    ErrorStringIndexExpr(Position),
    ErrorImportExpr(Position),
    ErrorInExpr(Position),
    ErrorLogicGuard(Position),
    ErrorFor(Position),
    ErrorVariableNotFound(String, Position),
    ErrorModuleNotFound(String, Position),
    ErrorDataRace(String, Position),
    ErrorAssignmentToUnknownLHS(Position),
    ErrorAssignmentToConstant(String, Position),
    ErrorMismatchOutputType(String, String, Position),
    ErrorDotExpr(String, Position),
    ErrorArithmetic(String, Position),
    ErrorTooManyOperations(Position),
    ErrorTooManyModules(Position),
    ErrorStackOverflow(Position),
    ErrorDataTooLarge(String, usize, usize, Position),
    ErrorTerminated(Position),
    ErrorRuntime(String, Position),
    ErrorLoopBreak(bool, Position),
    Return(Dynamic, Position),
}

Evaluation result.

All wrapped Position values represent the location in the script where the error occurs.

Currently, EvalAltResult is neither Send nor Sync. Turn on the sync feature to make it Send + Sync.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

ErrorParsing(ParseErrorType, Position)

Syntax error.

ErrorReadingScriptFile(PathBuf, Position, Error)

Error reading from a script file. Wrapped value is the path of the script file.

Never appears under the no_std feature.

ErrorFunctionNotFound(String, Position)

Call to an unknown function. Wrapped value is the signature of the function.

ErrorInFunctionCall(String, Box<EvalAltResult>, Position)

An error has occurred inside a called function. Wrapped values are the name of the function and the interior error.

ErrorUnboundThis(Position)

Access to this that is not bound.

ErrorBooleanArgMismatch(String, Position)

Non-boolean operand encountered for boolean operator. Wrapped value is the operator.

ErrorCharMismatch(Position)

Non-character value encountered where a character is required.

ErrorArrayBounds(usize, i32, Position)

Array access out-of-bounds. Wrapped values are the current number of elements in the array and the index number.

ErrorStringBounds(usize, i32, Position)

String indexing out-of-bounds. Wrapped values are the current number of characters in the string and the index number.

ErrorIndexingType(String, Position)

Trying to index into a type that is not an array, an object map, or a string, and has no indexer function defined.

ErrorNumericIndexExpr(Position)

Trying to index into an array or string with an index that is not i64.

ErrorStringIndexExpr(Position)

Trying to index into a map with an index that is not String.

ErrorImportExpr(Position)

Trying to import with an expression that is not String.

ErrorInExpr(Position)

Invalid arguments for in operator.

ErrorLogicGuard(Position)

The guard expression in an if or while statement does not return a boolean value.

ErrorFor(Position)

The for statement encounters a type that is not an iterator.

ErrorVariableNotFound(String, Position)

Usage of an unknown variable. Wrapped value is the name of the variable.

ErrorModuleNotFound(String, Position)

Usage of an unknown module. Wrapped value is the name of the module.

ErrorDataRace(String, Position)

Data race detected when accessing a variable. Wrapped value is the name of the variable.

ErrorAssignmentToUnknownLHS(Position)

Assignment to an inappropriate LHS (left-hand-side) expression.

ErrorAssignmentToConstant(String, Position)

Assignment to a constant variable.

ErrorMismatchOutputType(String, String, Position)

Returned type is not the same as the required output type. Wrapped values are the type requested and type of the actual result.

ErrorDotExpr(String, Position)

Inappropriate member access.

ErrorArithmetic(String, Position)

Arithmetic error encountered. Wrapped value is the error message.

ErrorTooManyOperations(Position)

Number of operations over maximum limit.

ErrorTooManyModules(Position)

Modules over maximum limit.

ErrorStackOverflow(Position)

Call stack over maximum limit.

ErrorDataTooLarge(String, usize, usize, Position)

Data value over maximum size limit. Wrapped values are the data type, maximum size and current size.

ErrorTerminated(Position)

The script is prematurely terminated.

ErrorRuntime(String, Position)

Run-time error encountered. Wrapped value is the error message.

ErrorLoopBreak(bool, Position)

Breaking out of loops - not an error if within a loop. The wrapped value, if true, means breaking clean out of the loop (i.e. a break statement). The wrapped value, if false, means breaking the current context (i.e. a continue statement).

Return(Dynamic, Position)

Not an error: Value returned from a script via the return keyword. Wrapped value is the result value.

Implementations

impl EvalAltResult

pub fn position(&self) -> Position

Get the Position of this error.

pub fn set_position(&mut self, new_position: Position)

Override the Position of this error.