Enum casbin::rhai::ParseErrorType

#[non_exhaustive]
pub enum ParseErrorType {

    UnexpectedEOF,
    BadInput(LexError),
    UnknownOperator(String),
    MissingToken(String, String),
    MissingSymbol(String),
    MalformedCallExpr(String),
    MalformedIndexExpr(String),
    MalformedInExpr(String),
    MalformedCapture(String),
    DuplicatedProperty(String),
    DuplicatedSwitchCase,
    DuplicatedVariable(String),
    WrongSwitchIntegerCase,
    WrongSwitchDefaultCase,
    WrongSwitchCaseCondition,
    PropertyExpected,
    VariableExpected,
    ForbiddenVariable(String),
    Reserved(String),
    MismatchedType(String, String),
    ExprExpected(String),
    WrongDocComment,
    WrongFnDefinition,
    FnDuplicatedDefinition(String, usize),
    FnMissingName,
    FnMissingParams(String),
    FnDuplicatedParam(String, String),
    FnMissingBody(String),
    WrongExport,
    AssignmentToConstant(String),
    AssignmentToInvalidLHS(String),
    VariableExists(String),
    VariableUndefined(String),
    ModuleUndefined(String),
    ExprTooDeep,
    LiteralTooLarge(String, usize),
    LoopBreak,

}

Error encountered when parsing a script.

Some errors never appear when certain features are turned on. They still exist so that the application can turn features on and off without going through massive code changes to remove/add back enum variants in match statements.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

UnexpectedEOF

The script ends prematurely.

BadInput(LexError)

Tuple Fields

0: LexError

Error in the script text. Wrapped value is the lex error.

UnknownOperator(String)

Tuple Fields

0: String

An unknown operator is encountered. Wrapped value is the operator.

MissingToken(String, String)

Tuple Fields

0: String

1: String

Expecting a particular token but not finding one. Wrapped values are the token and description.

MissingSymbol(String)

Tuple Fields

0: String

Expecting a particular symbol but not finding one. Wrapped value is the description.

MalformedCallExpr(String)

Tuple Fields

0: String

An expression in function call arguments () has syntax error. Wrapped value is the error description (if any).

MalformedIndexExpr(String)

Tuple Fields

0: String

An expression in indexing brackets [] has syntax error. Wrapped value is the error description (if any).

MalformedInExpr(String)

Tuple Fields

0: String

An expression in an in expression has syntax error. Wrapped value is the error description (if any).

MalformedCapture(String)

Tuple Fields

0: String

A capturing has syntax error. Wrapped value is the error description (if any).

DuplicatedProperty(String)

Tuple Fields

0: String

A map definition has duplicated property names. Wrapped value is the property name.

DuplicatedSwitchCase

A switch case is duplicated.

DuplicatedVariable(String)

Tuple Fields

0: String

A variable name is duplicated. Wrapped value is the variable name.

WrongSwitchIntegerCase

An integer case of a switch statement is in an appropriate place.

WrongSwitchDefaultCase

The default case of a switch statement is in an appropriate place.

WrongSwitchCaseCondition

The case condition of a switch statement is not appropriate.

PropertyExpected

Missing a property name for custom types and maps.

VariableExpected

Missing a variable name after the let, const, for or catch keywords.

ForbiddenVariable(String)

Tuple Fields

0: String

Forbidden variable name. Wrapped value is the variable name.

Reserved(String)

Tuple Fields

0: String

An identifier is a reserved symbol.

MismatchedType(String, String)

Tuple Fields

0: String

1: String

An expression is of the wrong type. Wrapped values are the type requested and type of the actual result.

ExprExpected(String)

Tuple Fields

0: String

Missing an expression. Wrapped value is the expression type.

WrongDocComment

Defining a doc-comment in an appropriate place (e.g. not at global level).

WrongFnDefinition

Defining a function fn in an appropriate place (e.g. inside another function).

FnDuplicatedDefinition(String, usize)

Tuple Fields

0: String

1: usize

Defining a function with a name that conflicts with an existing function. Wrapped values are the function name and number of parameters.

FnMissingName

Missing a function name after the fn keyword.

FnMissingParams(String)

Tuple Fields

0: String

A function definition is missing the parameters list. Wrapped value is the function name.

FnDuplicatedParam(String, String)

Tuple Fields

0: String

1: String

A function definition has duplicated parameters. Wrapped values are the function name and parameter name.

FnMissingBody(String)

Tuple Fields

0: String

A function definition is missing the body. Wrapped value is the function name.

WrongExport

Export statement not at global level.

AssignmentToConstant(String)

Tuple Fields

0: String

Assignment to an a constant variable. Wrapped value is the constant variable name.

AssignmentToInvalidLHS(String)

Tuple Fields

0: String

Assignment to an inappropriate LHS (left-hand-side) expression. Wrapped value is the error message (if any).

VariableExists(String)

Tuple Fields

0: String

A variable is already defined.

Only appears when variables shadowing is disabled.

VariableUndefined(String)

Tuple Fields

0: String

A variable is not found.

Only appears when strict variables mode is enabled.

ModuleUndefined(String)

Tuple Fields

0: String

An imported module is not found.

Only appears when strict variables mode is enabled.

ExprTooDeep

Expression exceeding the maximum levels of complexity.

LiteralTooLarge(String, usize)

Tuple Fields

0: String

1: usize

Literal exceeding the maximum size. Wrapped values are the data type name and the maximum size.

LoopBreak

Break statement not inside a loop.