Enum casbin::rhai::ParseErrorType

#[non_exhaustive]pub enum ParseErrorType {
    BadInput(String),
    UnexpectedEOF,
    UnknownOperator(String),
    MissingToken(String, String),
    MalformedCallExpr(String),
    MalformedIndexExpr(String),
    MalformedInExpr(String),
    MalformedCapture(String),
    DuplicatedProperty(String),
    ForbiddenConstantExpr(String),
    PropertyExpected,
    VariableExpected,
    Reserved(String),
    ExprExpected(String),
    WrongFnDefinition,
    FnMissingName,
    FnMissingParams(String),
    FnDuplicatedParam(String, String),
    FnMissingBody(String),
    DuplicatedExport(String),
    WrongExport,
    AssignmentToCopy,
    AssignmentToConstant(String),
    ExprTooDeep,
    LiteralTooLarge(String, usize),
    LoopBreak,
}

Type of error encountered when parsing a script.

Some errors never appear when certain features are turned on. They still exist so that the application can turn features on and off without going through massive code changes to remove/add back enum variants in match statements.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

BadInput(String)

Error in the script text. Wrapped value is the error message.

UnexpectedEOF

The script ends prematurely.

UnknownOperator(String)

An unknown operator is encountered. Wrapped value is the operator.

MissingToken(String, String)

Expecting a particular token but not finding one. Wrapped values are the token and description.

MalformedCallExpr(String)

An expression in function call arguments () has syntax error. Wrapped value is the error description (if any).

MalformedIndexExpr(String)

An expression in indexing brackets [] has syntax error. Wrapped value is the error description (if any).

Never appears under the no_index feature.

MalformedInExpr(String)

An expression in an in expression has syntax error. Wrapped value is the error description (if any).

Never appears under the no_object and no_index features combination.

MalformedCapture(String)

A capturing has syntax error. Wrapped value is the error description (if any).

Never appears under the no_closure feature.

DuplicatedProperty(String)

A map definition has duplicated property names. Wrapped value is the property name.

Never appears under the no_object feature.

ForbiddenConstantExpr(String)

Invalid expression assigned to constant. Wrapped value is the name of the constant.

PropertyExpected

Missing a property name for custom types and maps.

Never appears under the no_object feature.

VariableExpected

Missing a variable name after the let, const or for keywords.

Reserved(String)

An identifier is a reserved keyword.

ExprExpected(String)

Missing an expression. Wrapped value is the expression type.

WrongFnDefinition

Defining a function fn in an appropriate place (e.g. inside another function).

Never appears under the no_function feature.

FnMissingName

Missing a function name after the fn keyword.

Never appears under the no_function feature.

FnMissingParams(String)

A function definition is missing the parameters list. Wrapped value is the function name.

Never appears under the no_function feature.

FnDuplicatedParam(String, String)

A function definition has duplicated parameters. Wrapped values are the function name and parameter name.

Never appears under the no_function feature.

FnMissingBody(String)

A function definition is missing the body. Wrapped value is the function name.

Never appears under the no_function feature.

DuplicatedExport(String)

An export statement has duplicated names.

Never appears under the no_module feature.

WrongExport

Export statement not at global level.

Never appears under the no_module feature.

AssignmentToCopy

Assignment to a copy of a value.

AssignmentToConstant(String)

Assignment to an a constant variable. Wrapped value is the constant variable name.

ExprTooDeep

Expression exceeding the maximum levels of complexity.

Never appears under the unchecked feature.

LiteralTooLarge(String, usize)

Literal exceeding the maximum size. Wrapped values are the data type name and the maximum size.

Never appears under the unchecked feature.

LoopBreak

Break statement not inside a loop.