Enum casbin::rhai::plugin::CallableFunction

pub enum CallableFunction {
    Pure(Arc<dyn Fn(NativeCallContext<'_>, &mut [&mut Dynamic]) + Sync + Send + 'static>),
    Method(Arc<dyn Fn(NativeCallContext<'_>, &mut [&mut Dynamic]) + Sync + Send + 'static>),
    Iterator(fn(Dynamic) -> Box<dyn Iterator<Item = Dynamic> + 'static, Global>),
    Plugin(Arc<dyn PluginFunction + Sync + Send + 'static>),
}

A type encapsulating a function callable by Rhai.

Variants

Pure(Arc<dyn Fn(NativeCallContext<'_>, &mut [&mut Dynamic]) + Sync + Send + 'static>)

A pure native Rust function with all arguments passed by value.

Tuple Fields of Pure

0: Arc<dyn Fn(NativeCallContext<'_>, &mut [&mut Dynamic]) + Sync + Send + 'static>

Method(Arc<dyn Fn(NativeCallContext<'_>, &mut [&mut Dynamic]) + Sync + Send + 'static>)

A native Rust object method with the first argument passed by reference, and the rest passed by value.

Tuple Fields of Method

0: Arc<dyn Fn(NativeCallContext<'_>, &mut [&mut Dynamic]) + Sync + Send + 'static>

Iterator(fn(Dynamic) -> Box<dyn Iterator<Item = Dynamic> + 'static, Global>)

An iterator function.

Tuple Fields of Iterator

0: fn(Dynamic) -> Box<dyn Iterator<Item = Dynamic> + 'static, Global>

Plugin(Arc<dyn PluginFunction + Sync + Send + 'static>)

A plugin function,

Tuple Fields of Plugin

0: Arc<dyn PluginFunction + Sync + Send + 'static>

Implementations

impl CallableFunction

pub fn is_pure(&self) -> bool

Is this a pure native Rust function?

pub fn is_method(&self) -> bool

Is this a native Rust method function?

pub const fn is_iter(&self) -> bool

Is this an iterator function?

pub const fn is_script(&self) -> bool

Is this a Rhai-scripted function?

pub const fn is_plugin_fn(&self) -> bool

Is this a plugin function?

pub const fn is_native(&self) -> bool

Is this a native Rust function?

pub fn access(&self) -> FnAccess

Get the access mode.

pub fn get_native_fn(
&self
) -> Option<&Arc<dyn Fn(NativeCallContext<'_>, &mut [&mut Dynamic]) + Sync + Send + 'static>>

Get a shared reference to a native Rust function.

pub fn get_iter_fn(
&self
) -> Option<fn(Dynamic) -> Box<dyn Iterator<Item = Dynamic> + 'static, Global>>

Get a reference to an iterator function.

pub fn get_plugin_fn(
&self
) -> Option<&Arc<dyn PluginFunction + Sync + Send + 'static>>

Get a shared reference to a plugin function.

pub fn from_pure(
func: Box<dyn Fn(NativeCallContext<'_>, &mut [&mut Dynamic]) + Sync + Send + 'static, Global>
) -> CallableFunction

Create a new CallableFunction::Pure.

pub fn from_method(
func: Box<dyn Fn(NativeCallContext<'_>, &mut [&mut Dynamic]) + Sync + Send + 'static, Global>
) -> CallableFunction

Create a new CallableFunction::Method.

pub fn from_plugin(
func: impl PluginFunction + SendSync + 'static
) -> CallableFunction

Create a new CallableFunction::Plugin.