Enum casbin::rhai::plugin::CallableFunction

pub enum CallableFunction {
    Pure(Arc<dyn Fn(NativeCallContext<'_, '_, '_, '_, '_>, &mut [&mut Dynamic]) + 'static + Sync + Send>),
    Method(Arc<dyn Fn(NativeCallContext<'_, '_, '_, '_, '_>, &mut [&mut Dynamic]) + 'static + Sync + Send>),
    Iterator(fn(Dynamic) -> Box<dyn Iterator<Item = Dynamic> + 'static, Global>),
    Plugin(Arc<dyn PluginFunction + 'static + Sync + Send>),
}

A type encapsulating a function callable by Rhai.

Variants

Pure(Arc<dyn Fn(NativeCallContext<'_, '_, '_, '_, '_>, &mut [&mut Dynamic]) + 'static + Sync + Send>)

A pure native Rust function with all arguments passed by value.

Method(Arc<dyn Fn(NativeCallContext<'_, '_, '_, '_, '_>, &mut [&mut Dynamic]) + 'static + Sync + Send>)

A native Rust object method with the first argument passed by reference, and the rest passed by value.

Iterator(fn(Dynamic) -> Box<dyn Iterator<Item = Dynamic> + 'static, Global>)

An iterator function.

Plugin(Arc<dyn PluginFunction + 'static + Sync + Send>)

A plugin function,

Implementations

impl CallableFunction

pub fn is_pure(&self) -> bool

Is this a pure native Rust function?

pub fn is_method(&self) -> bool

Is this a native Rust method function?

pub fn is_iter(&self) -> bool

Is this an iterator function?

pub fn is_script(&self) -> bool

Is this a Rhai-scripted function?

pub fn is_plugin_fn(&self) -> bool

Is this a plugin function?

pub fn is_native(&self) -> bool

Is this a native Rust function?

pub fn access(&self) -> FnAccess

Get the access mode.

pub fn get_native_fn(
&self
) -> &(dyn Fn(NativeCallContext<'_, '_, '_, '_, '_>, &mut [&mut Dynamic]) + 'static + Sync + Send)

Get a shared reference to a native Rust function.

Panics

Panics if the CallableFunction is not Pure or Method.

pub fn get_iter_fn(
&self
) -> fn(Dynamic) -> Box<dyn Iterator<Item = Dynamic> + 'static, Global>

Get a reference to an iterator function.

Panics

Panics if the CallableFunction is not Iterator.

pub fn get_plugin_fn(
&'s self
) -> &'s (dyn PluginFunction + 'static + Sync + Send)

Get a shared reference to a plugin function.

Panics

Panics if the CallableFunction is not Plugin.

pub fn from_pure(
func: Box<dyn Fn(NativeCallContext<'_, '_, '_, '_, '_>, &mut [&mut Dynamic]) + 'static + Sync + Send, Global>
) -> CallableFunction

Create a new CallableFunction::Pure.

pub fn from_method(
func: Box<dyn Fn(NativeCallContext<'_, '_, '_, '_, '_>, &mut [&mut Dynamic]) + 'static + Sync + Send, Global>
) -> CallableFunction

Create a new CallableFunction::Method.

pub fn from_plugin(
func: impl SendSync + PluginFunction + 'static
) -> CallableFunction

Create a new CallableFunction::Plugin.