Struct casbin::rhai::plugin::FuncRegistration

pub struct FuncRegistration { /* private fields */ }

Type for fine-tuned module function registration.

Implementations

impl FuncRegistration

pub fn new(name: impl Into<SmartString<LazyCompact>>) -> FuncRegistration

Create a new FuncRegistration.

Defaults

Accessibility: The function namespace is FnNamespace::Internal.

Purity: The function is assumed to be pure unless it is a property setter or an index setter.

Volatility: The function is assumed to be non-volatile – i.e. it guarantees the same result for the same input(s).

Metadata: No metadata for the function is registered.

let mut module = Module::new();

fn inc(x: i64) -> i64 { x + 1 }

let f = FuncRegistration::new("inc")
    .with_namespace(FnNamespace::Global)
    .set_into_module(&mut module, inc);

let hash = f.hash;

assert!(module.contains_fn(hash));

pub fn new_getter(prop: impl AsRef<str>) -> FuncRegistration

Create a new FuncRegistration for a property getter.

Not available under no_object.

Defaults

Accessibility: The function namespace is FnNamespace::Global.

Purity: The function is assumed to be pure.

Volatility: The function is assumed to be non-volatile – i.e. it guarantees the same result for the same input(s).

Metadata: No metadata for the function is registered.

pub fn new_setter(prop: impl AsRef<str>) -> FuncRegistration

Create a new FuncRegistration for a property setter.

Not available under no_object.

Defaults

Accessibility: The function namespace is FnNamespace::Global.

Purity: The function is assumed to be no-pure.

Volatility: The function is assumed to be non-volatile – i.e. it guarantees the same result for the same input(s).

Metadata: No metadata for the function is registered.

pub fn new_index_getter() -> FuncRegistration

Create a new FuncRegistration for an index getter.

Not available under both no_index and no_object.

Defaults

Accessibility: The function namespace is FnNamespace::Global.

Purity: The function is assumed to be pure.

Volatility: The function is assumed to be non-volatile – i.e. it guarantees the same result for the same input(s).

Metadata: No metadata for the function is registered.

pub fn new_index_setter() -> FuncRegistration

Create a new FuncRegistration for an index setter.

Not available under both no_index and no_object.

Defaults

Accessibility: The function namespace is FnNamespace::Global.

Purity: The function is assumed to be no-pure.

Volatility: The function is assumed to be non-volatile – i.e. it guarantees the same result for the same input(s).

Metadata: No metadata for the function is registered.

pub fn with_namespace(self, namespace: FnNamespace) -> FuncRegistration

Set the namespace of the function.

pub fn with_purity(self, pure: bool) -> FuncRegistration

Set whether the function is pure. A pure function has no side effects.

pub fn with_volatility(self, volatile: bool) -> FuncRegistration

Set whether the function is volatile. A volatile function does not guarantee the same result for the same input(s).

pub fn register_into_engine<A, const N: usize, const X: bool, R, const F: bool, FUNC>( self, engine: &mut Engine, func: FUNC ) -> &FuncMetadatawhere A: 'static, R: Variant + Clone, FUNC: RhaiNativeFunc<A, N, X, R, F> + SendSync + 'static,

Register the function into the specified Engine.

pub fn set_into_module<A, const N: usize, const X: bool, R, const F: bool, FUNC>( self, module: &mut Module, func: FUNC ) -> &FuncMetadatawhere A: 'static, R: Variant + Clone, FUNC: RhaiNativeFunc<A, N, X, R, F> + SendSync + 'static,

Register the function into the specified Module.

pub fn set_into_module_raw( self, module: &mut Module, arg_types: impl AsRef<[TypeId]>, func: RhaiFunc ) -> &FuncMetadata

Register the function into the specified Module.

WARNING - Low Level API

This function is very low level. It takes a list of TypeId’s indicating the actual types of the parameters.

Panics

Panics if the type of the first parameter is Array, Map, String, ImmutableString, &str or INT and the function name indicates that it is an index getter or setter.

Indexers for arrays, object maps, strings and integers cannot be registered.