Trait casbin::rhai::plugin::PluginFunction

pub trait PluginFunction {
    // Required methods
    fn call(
        &self,
        context: Option<NativeCallContext<'_>>,
        args: &mut [&mut Dynamic]
    ) -> Result<Dynamic, Box<EvalAltResult, Global>>;
    fn is_method_call(&self) -> bool;
    fn has_context(&self) -> bool;

    // Provided method
    fn is_pure(&self) -> bool { ... }
}

Trait implemented by a plugin function.

This trait should not be used directly. Use the #[export_module] and #[export_fn] procedural attributes instead.

Required Methods

fn call( &self, context: Option<NativeCallContext<'_>>, args: &mut [&mut Dynamic] ) -> Result<Dynamic, Box<EvalAltResult, Global>>

Call the plugin function with the arguments provided.

fn is_method_call(&self) -> bool

Is this plugin function a method?

fn has_context(&self) -> bool

Does this plugin function contain a NativeCallContext parameter?

Provided Methods

fn is_pure(&self) -> bool

Is this plugin function pure?

This defaults to true such that any old implementation that has constant-checking code inside the function itself will continue to work.