Trait casbin::rhai::plugin::PluginFunction

pub trait PluginFunction {
    fn call(
        &self, 
        context: NativeCallContext<'_, '_, '_, '_>, 
        args: &mut [&mut Dynamic]
    ) -> Result<Dynamic, Box<EvalAltResult>>;

    fn is_method_call(&self) -> bool;

    fn is_variadic(&self) -> bool;

    fn clone_boxed(&self) -> Box<dyn PluginFunction + 'static>;

    fn input_names(&self) -> Box<[&'static str]>;

    fn input_types(&self) -> Box<[TypeId]>;

    fn return_type(&self) -> &'static str;
}

Trait implemented by a plugin function. This trait should not be used directly.

Use the #[export_module] and #[export_fn] procedural attributes instead.

Required methods

fn call(
&self,
context: NativeCallContext<'_, '_, '_, '_>,
args: &mut [&mut Dynamic]
) -> Result<Dynamic, Box<EvalAltResult>>

Call the plugin function with the arguments provided.

fn is_method_call(&self) -> bool

Is this plugin function a method?

fn is_variadic(&self) -> bool

Is this plugin function variadic?

fn clone_boxed(&self) -> Box<dyn PluginFunction + 'static>

Convert a plugin function into a boxed trait object.

fn input_names(&self) -> Box<[&'static str]>

Return a boxed slice of the names of the function's parameters.

fn input_types(&self) -> Box<[TypeId]>

Return a boxed slice of type ID's of the function's parameters.

fn return_type(&self) -> &'static str

Return a string slice of the function's return type.

Loading content...