Struct casbin::rhai::EvalContext

pub struct EvalContext<'a, 's, 'ps, 'g, 'c, 't> { /* private fields */ }

Context of a script evaluation process.

Implementations

impl<'a, 's, 'ps, 'g, 'c, 't> EvalContext<'a, 's, 'ps, 'g, 'c, 't>

pub fn new(
engine: &'a Engine,
global: &'g mut GlobalRuntimeState,
caches: &'c mut Caches,
scope: &'s mut Scope<'ps, _>,
this_ptr: Option<&'t mut Dynamic>
) -> EvalContext<'a, 's, 'ps, 'g, 'c, 't>

Create a new EvalContext.

pub const fn engine(&self) -> &'a Engine

The current Engine.

pub fn source(&self) -> Option<&str>

The current source.

pub const fn scope(&self) -> &Scope<'ps, _>

The current Scope.

pub fn scope_mut(&mut self) -> &mut Scope<'ps, _>

Get a mutable reference to the current Scope.

pub const fn tag(&self) -> &Dynamic

Custom state kept in a Dynamic.

pub fn tag_mut(&mut self) -> &mut Dynamic

Mutable reference to the custom state kept in a Dynamic.

pub fn this_ptr(&self) -> Option<&Dynamic>

The current bound this pointer, if any.

pub fn this_ptr_mut(&mut self) -> Option<&mut Dynamic>

Mutable reference to the current bound this pointer, if any.

pub const fn call_level(&self) -> usize

The current nesting level of function calls.

pub fn eval_expression_tree(
&mut self,
expr: &Expression<'_>
) -> Result<Dynamic, Box<EvalAltResult, Global>>

Evaluate an expression tree within this evaluation context.

WARNING - Low Level API

This function is very low level. It evaluates an expression from an AST.

pub fn eval_expression_tree_raw(
&mut self,
expr: &Expression<'_>,
rewind_scope: bool
) -> Result<Dynamic, Box<EvalAltResult, Global>>

👎Deprecated: This API is NOT deprecated, but it is considered volatile and may change in the future.

Evaluate an expression tree within this evaluation context.

The following option is available:

whether to rewind the Scope after evaluation if the expression is a [StmtBlock][crate::ast::StmtBlock]

WARNING - Unstable API

This API is volatile and may change in the future.

WARNING - Low Level API

This function is extremely low level. It evaluates an expression from an AST.