Struct casbin::rhai::EvalContext

pub struct EvalContext<'a, 'x, 'px, 'm, 's, 'b, 't, 'pt> { /* fields omitted */ }

Context of a script evaluation process.

Implementations

impl<'_, '_, '_, '_, '_, '_, '_, '_> EvalContext<'_, '_, '_, '_, '_, '_, '_, '_>

pub fn eval_expression_tree(
&mut self,
expr: &Expression<'_>
) -> Result<Dynamic, Box<EvalAltResult, Global>>

Evaluate an expression tree.

WARNING - Low Level API

This function is very low level. It evaluates an expression from an AST.

impl<'x, 'px, 'pt, '_, '_, '_, '_, '_> EvalContext<'_, 'x, 'px, '_, '_, '_, '_, 'pt>

#[must_use]
pub const fn engine(&self) -> &Engine

The current Engine.

#[must_use]
pub fn source(&self) -> Option<&str>

The current source.

#[must_use]
pub const fn scope(&self) -> &Scope<'px>

The current Scope.

#[must_use]
pub fn scope_mut(&mut self) -> &mut &'x mut Scope<'px>

Mutable reference to the current Scope.

pub fn iter_namespaces(&self) -> impl Iterator<Item = &Module>

Get an iterator over the namespaces containing definition of all script-defined functions.

#[must_use]
pub fn this_ptr(&self) -> Option<&Dynamic>

The current bound this pointer, if any.

#[must_use]
pub fn this_ptr_mut(&mut self) -> Option<&mut &'pt mut Dynamic>

Mutable reference to the current bound this pointer, if any.

#[must_use]
pub const fn call_level(&self) -> usize

The current nesting level of function calls.