Struct casbin::rhai::FnPtr

pub struct FnPtr { /* private fields */ }

A general function pointer, which may carry additional (i.e. curried) argument values to be passed onto a function during a call.

Implementations

impl FnPtr

pub fn num_curried(&self) -> usize

๐Deprecated since 1.8.0: use curry().len() instead

Get the number of curried arguments.

Deprecated

This method is deprecated. Use curry().len() instead.

This method will be removed in the next major version.

pub fn call_dynamic( &self, context: &NativeCallContext<'_>, this_ptr: Option<&mut Dynamic>, arg_values: impl AsMut<[Dynamic]> ) -> Result<Dynamic, Box<EvalAltResult, Global>>

๐Deprecated since 1.3.0: use call_within_context or call_raw instead

Call the function pointer with curried arguments (if any). The function may be script-defined (not available under no_function) or native Rust.

This method is intended for calling a function pointer that is passed into a native Rust function as an argument. Therefore, the AST is NOT evaluated before calling the function.

Deprecated

This method is deprecated. Use call_within_context or call_raw instead.

This method will be removed in the next major version.

impl FnPtr

pub fn new( name: impl Into<ImmutableString> ) -> Result<FnPtr, Box<EvalAltResult, Global>>

Create a new function pointer.

pub fn fn_name(&self) -> &str

Get the name of the function.

pub fn curry(&self) -> &[Dynamic]

Get the curried arguments.

pub fn iter_curry(&self) -> impl Iterator<Item = &Dynamic>

Iterate the curried arguments.

pub fn iter_curry_mut(&mut self) -> impl Iterator<Item = &mut Dynamic>

Mutably-iterate the curried arguments.

pub fn add_curry(&mut self, value: Dynamic) -> &mut FnPtr

Add a new curried argument.

pub fn set_curry( &mut self, values: impl IntoIterator<Item = Dynamic> ) -> &mut FnPtr

Set curried arguments to the function pointer.

pub fn is_curried(&self) -> bool

Is the function pointer curried?

pub fn call<T>( &self, engine: &Engine, ast: &AST, args: impl FuncArgs ) -> Result<T, Box<EvalAltResult, Global>>where T: Variant + Clone,

Call the function pointer with curried arguments (if any). The function may be script-defined (not available under no_function) or native Rust.

This method is intended for calling a function pointer directly, possibly on another Engine. Therefore, the AST is NOT evaluated before calling the function.

Example

use rhai::{Engine, FnPtr};

let engine = Engine::new();

let ast = engine.compile("fn foo(x, y) { len(x) + y }")?;

let mut fn_ptr = FnPtr::new("foo")?;

// Curry values into the function pointer
fn_ptr.set_curry(vec!["abc".into()]);

// Values are only needed for non-curried parameters
let result: i64 = fn_ptr.call(&engine, &ast, ( 39_i64, ) )?;

assert_eq!(result, 42);

pub fn call_within_context<T>( &self, context: &NativeCallContext<'_>, args: impl FuncArgs ) -> Result<T, Box<EvalAltResult, Global>>where T: Variant + Clone,

Call the function pointer with curried arguments (if any). The function may be script-defined (not available under no_function) or native Rust.

This method is intended for calling a function pointer that is passed into a native Rust function as an argument. Therefore, the AST is NOT evaluated before calling the function.

pub fn call_raw( &self, context: &NativeCallContext<'_>, this_ptr: Option<&mut Dynamic>, arg_values: impl AsMut<[Dynamic]> ) -> Result<Dynamic, Box<EvalAltResult, Global>>

Call the function pointer with curried arguments (if any). The function may be script-defined (not available under no_function) or native Rust.

This method is intended for calling a function pointer that is passed into a native Rust function as an argument. Therefore, the AST is NOT evaluated before calling the function.

WARNING - Low Level API

This function is very low level.

Arguments

All the arguments are consumed, meaning that theyโre replaced by (). This is to avoid unnecessarily cloning the arguments.

Do not use the arguments after this call. If they are needed afterwards, clone them before calling this function.