Struct casbin::rhai::Scope

pub struct Scope<'a>(_);

Type containing information about the current scope. Useful for keeping state between Engine evaluation runs.

Example

use rhai::{Engine, Scope};

let engine = Engine::new();
let mut my_scope = Scope::new();

my_scope.push("z", 40_i64);

engine.eval_with_scope::<()>(&mut my_scope, "let x = z + 1; z = 0;")?;

assert_eq!(engine.eval_with_scope::<i64>(&mut my_scope, "x + 1")?, 42);

assert_eq!(my_scope.get_value::<i64>("x").unwrap(), 41);
assert_eq!(my_scope.get_value::<i64>("z").unwrap(), 0);

When searching for entries, newly-added entries are found before similarly-named but older entries, allowing for automatic shadowing.

Currently, Scope is neither Send nor Sync. Turn on the sync feature to make it Send + Sync.

Implementations

impl<'a> Scope<'a>

pub fn new() -> Scope<'a>

Create a new Scope.

Examples

use rhai::Scope;

let mut my_scope = Scope::new();

my_scope.push("x", 42_i64);
assert_eq!(my_scope.get_value::<i64>("x").unwrap(), 42);

pub fn clear(&mut self) -> &mut Scope<'a>

Empty the Scope.

Examples

use rhai::Scope;

let mut my_scope = Scope::new();

my_scope.push("x", 42_i64);
assert!(my_scope.contains("x"));
assert_eq!(my_scope.len(), 1);
assert!(!my_scope.is_empty());

my_scope.clear();
assert!(!my_scope.contains("x"));
assert_eq!(my_scope.len(), 0);
assert!(my_scope.is_empty());

pub fn len(&self) -> usize

Get the number of entries inside the Scope.

Examples

use rhai::Scope;

let mut my_scope = Scope::new();
assert_eq!(my_scope.len(), 0);

my_scope.push("x", 42_i64);
assert_eq!(my_scope.len(), 1);

pub fn is_empty(&self) -> bool

Is the Scope empty?

Examples

use rhai::Scope;

let mut my_scope = Scope::new();
assert!(my_scope.is_empty());

my_scope.push("x", 42_i64);
assert!(!my_scope.is_empty());

pub fn push<K, T>(&mut self, name: K, value: T) -> &mut Scope<'a> where
K: Into<Cow<'a, str>>,
T: Variant + Clone,

Add (push) a new entry to the Scope.

Examples

use rhai::Scope;

let mut my_scope = Scope::new();

my_scope.push("x", 42_i64);
assert_eq!(my_scope.get_value::<i64>("x").unwrap(), 42);

pub fn push_dynamic<K>(&mut self, name: K, value: Dynamic) -> &mut Scope<'a> where
K: Into<Cow<'a, str>>,

Add (push) a new Dynamic entry to the Scope.

Examples

use rhai::{Dynamic,  Scope};

let mut my_scope = Scope::new();

my_scope.push_dynamic("x", Dynamic::from(42_i64));
assert_eq!(my_scope.get_value::<i64>("x").unwrap(), 42);

pub fn push_constant<K, T>(&mut self, name: K, value: T) -> &mut Scope<'a> where
K: Into<Cow<'a, str>>,
T: Variant + Clone,

Add (push) a new constant to the Scope.

Constants are immutable and cannot be assigned to. Their values never change. Constants propagation is a technique used to optimize an AST.

However, in order to be used for optimization, constants must be in one of the recognized types: INT (default to i64, i32 if only_i32), f64, String, char and bool.

Examples

use rhai::Scope;

let mut my_scope = Scope::new();

my_scope.push_constant("x", 42_i64);
assert_eq!(my_scope.get_value::<i64>("x").unwrap(), 42);

pub fn push_constant_dynamic<K>(
&mut self,
name: K,
value: Dynamic
) -> &mut Scope<'a> where
K: Into<Cow<'a, str>>,

Add (push) a new constant with a Dynamic value to the Scope.

Constants are immutable and cannot be assigned to. Their values never change. Constants propagation is a technique used to optimize an AST.

However, in order to be used for optimization, the Dynamic value must be in one of the recognized types: INT (default to i64, i32 if only_i32), f64, String, char and bool.

Examples

use rhai::{Dynamic, Scope};

let mut my_scope = Scope::new();

my_scope.push_constant_dynamic("x", Dynamic::from(42_i64));
assert_eq!(my_scope.get_value::<i64>("x").unwrap(), 42);

pub fn rewind(&mut self, size: usize) -> &mut Scope<'a>

Truncate (rewind) the Scope to a previous size.

Examples

use rhai::Scope;

let mut my_scope = Scope::new();

my_scope.push("x", 42_i64);
my_scope.push("y", 123_i64);
assert!(my_scope.contains("x"));
assert!(my_scope.contains("y"));
assert_eq!(my_scope.len(), 2);

my_scope.rewind(1);
assert!(my_scope.contains("x"));
assert!(!my_scope.contains("y"));
assert_eq!(my_scope.len(), 1);

my_scope.rewind(0);
assert!(!my_scope.contains("x"));
assert!(!my_scope.contains("y"));
assert_eq!(my_scope.len(), 0);
assert!(my_scope.is_empty());

pub fn contains(&self, name: &str) -> bool

Does the scope contain the entry?

Examples

use rhai::Scope;

let mut my_scope = Scope::new();

my_scope.push("x", 42_i64);
assert!(my_scope.contains("x"));
assert!(!my_scope.contains("y"));

pub fn get_value<T>(&self, name: &str) -> Option<T> where
T: Variant + Clone,

Get the value of an entry in the Scope, starting from the last.

Examples

use rhai::Scope;

let mut my_scope = Scope::new();

my_scope.push("x", 42_i64);
assert_eq!(my_scope.get_value::<i64>("x").unwrap(), 42);

pub fn set_value<T>(&mut self, name: &'a str, value: T) -> &mut Scope<'a> where
T: Variant + Clone,

Update the value of the named entry. Search starts backwards from the last, and only the first entry matching the specified name is updated. If no entry matching the specified name is found, a new one is added.

Panics

Panics when trying to update the value of a constant.

Examples

use rhai::Scope;

let mut my_scope = Scope::new();

my_scope.push("x", 42_i64);
assert_eq!(my_scope.get_value::<i64>("x").unwrap(), 42);

my_scope.set_value("x", 0_i64);
assert_eq!(my_scope.get_value::<i64>("x").unwrap(), 0);

pub fn iter(&self) -> impl Iterator<Item = (&str, &Dynamic)>

Get an iterator to entries in the Scope.

Examples

use rhai::{Dynamic, Scope};

let mut my_scope = Scope::new();

my_scope.push("x", 42_i64);
my_scope.push("foo", "hello".to_string());

let mut iter = my_scope.iter();

let (name, value) = iter.next().unwrap();
assert_eq!(name, "x");
assert_eq!(value.clone().cast::<i64>(), 42);

let (name, value) = iter.next().unwrap();
assert_eq!(name, "foo");
assert_eq!(value.clone().cast::<String>(), "hello");