Struct casbin::rhai::Scope

pub struct Scope<'a> { /* fields omitted */ }

Type containing information about the current scope. Useful for keeping state between Engine evaluation runs.

Thread Safety

Currently, Scope is neither Send nor Sync. Turn on the sync feature to make it Send + Sync.

Example

use rhai::{Engine, Scope};

let engine = Engine::new();
let mut my_scope = Scope::new();

my_scope.push("z", 40_i64);

engine.eval_with_scope::<()>(&mut my_scope, "let x = z + 1; z = 0;")?;

assert_eq!(engine.eval_with_scope::<i64>(&mut my_scope, "x + 1")?, 42);

assert_eq!(my_scope.get_value::<i64>("x").unwrap(), 41);
assert_eq!(my_scope.get_value::<i64>("z").unwrap(), 0);

When searching for entries, newly-added entries are found before similarly-named but older entries, allowing for automatic shadowing.

Implementations

impl<'a> Scope<'a>

pub fn new() -> Scope<'a>

Create a new Scope.

Example

use rhai::Scope;

let mut my_scope = Scope::new();

my_scope.push("x", 42_i64);
assert_eq!(my_scope.get_value::<i64>("x").unwrap(), 42);

pub fn clear(&mut self) -> &mut Scope<'a>

Empty the Scope.

Example

use rhai::Scope;

let mut my_scope = Scope::new();

my_scope.push("x", 42_i64);
assert!(my_scope.contains("x"));
assert_eq!(my_scope.len(), 1);
assert!(!my_scope.is_empty());

my_scope.clear();
assert!(!my_scope.contains("x"));
assert_eq!(my_scope.len(), 0);
assert!(my_scope.is_empty());

pub fn len(&self) -> usize

Get the number of entries inside the Scope.

Example

use rhai::Scope;

let mut my_scope = Scope::new();
assert_eq!(my_scope.len(), 0);

my_scope.push("x", 42_i64);
assert_eq!(my_scope.len(), 1);

pub fn is_empty(&self) -> bool

Is the Scope empty?

Example

use rhai::Scope;

let mut my_scope = Scope::new();
assert!(my_scope.is_empty());

my_scope.push("x", 42_i64);
assert!(!my_scope.is_empty());

pub fn push(
&mut self,
name: impl Into<Cow<'a, str>>,
value: impl Variant + Clone
) -> &mut Scope<'a>

Add (push) a new entry to the Scope.

Example

use rhai::Scope;

let mut my_scope = Scope::new();

my_scope.push("x", 42_i64);
assert_eq!(my_scope.get_value::<i64>("x").unwrap(), 42);

pub fn push_dynamic(
&mut self,
name: impl Into<Cow<'a, str>>,
value: Dynamic
) -> &mut Scope<'a>

Add (push) a new Dynamic entry to the Scope.

Example

use rhai::{Dynamic,  Scope};

let mut my_scope = Scope::new();

my_scope.push_dynamic("x", Dynamic::from(42_i64));
assert_eq!(my_scope.get_value::<i64>("x").unwrap(), 42);

pub fn push_constant(
&mut self,
name: impl Into<Cow<'a, str>>,
value: impl Variant + Clone
) -> &mut Scope<'a>

Add (push) a new constant to the Scope.

Constants are immutable and cannot be assigned to. Their values never change. Constants propagation is a technique used to optimize an AST.

Example

use rhai::Scope;

let mut my_scope = Scope::new();

my_scope.push_constant("x", 42_i64);
assert_eq!(my_scope.get_value::<i64>("x").unwrap(), 42);

pub fn push_constant_dynamic(
&mut self,
name: impl Into<Cow<'a, str>>,
value: Dynamic
) -> &mut Scope<'a>

Add (push) a new constant with a Dynamic value to the Scope.

Constants are immutable and cannot be assigned to. Their values never change. Constants propagation is a technique used to optimize an AST.

Example

use rhai::{Dynamic, Scope};

let mut my_scope = Scope::new();

my_scope.push_constant_dynamic("x", Dynamic::from(42_i64));
assert_eq!(my_scope.get_value::<i64>("x").unwrap(), 42);

pub fn rewind(&mut self, size: usize) -> &mut Scope<'a>

Truncate (rewind) the Scope to a previous size.

Example

use rhai::Scope;

let mut my_scope = Scope::new();

my_scope.push("x", 42_i64);
my_scope.push("y", 123_i64);
assert!(my_scope.contains("x"));
assert!(my_scope.contains("y"));
assert_eq!(my_scope.len(), 2);

my_scope.rewind(1);
assert!(my_scope.contains("x"));
assert!(!my_scope.contains("y"));
assert_eq!(my_scope.len(), 1);

my_scope.rewind(0);
assert!(!my_scope.contains("x"));
assert!(!my_scope.contains("y"));
assert_eq!(my_scope.len(), 0);
assert!(my_scope.is_empty());

pub fn contains(&self, name: &str) -> bool

Does the Scope contain the entry?

Example

use rhai::Scope;

let mut my_scope = Scope::new();

my_scope.push("x", 42_i64);
assert!(my_scope.contains("x"));
assert!(!my_scope.contains("y"));

pub fn get_value<T>(&self, name: &str) -> Option<T> where
T: Variant + Clone,

Get the value of an entry in the Scope, starting from the last.

Example

use rhai::Scope;

let mut my_scope = Scope::new();

my_scope.push("x", 42_i64);
assert_eq!(my_scope.get_value::<i64>("x").unwrap(), 42);

pub fn set_value(
&mut self,
name: &'a str,
value: impl Variant + Clone
) -> &mut Scope<'a>

Update the value of the named entry in the Scope.

Search starts backwards from the last, and only the first entry matching the specified name is updated. If no entry matching the specified name is found, a new one is added.

Panics

Panics when trying to update the value of a constant.

Example

use rhai::Scope;

let mut my_scope = Scope::new();

my_scope.push("x", 42_i64);
assert_eq!(my_scope.get_value::<i64>("x").unwrap(), 42);

my_scope.set_value("x", 0_i64);
assert_eq!(my_scope.get_value::<i64>("x").unwrap(), 0);

pub fn get_mut(&mut self, name: &str) -> Option<&mut Dynamic>

Get a mutable reference to an entry in the Scope.

If the entry by the specified name is not found, of if it is read-only, None is returned.

Example

use rhai::Scope;

let mut my_scope = Scope::new();

my_scope.push("x", 42_i64);
assert_eq!(my_scope.get_value::<i64>("x").unwrap(), 42);

let ptr = my_scope.get_mut("x").unwrap();
*ptr = 123_i64.into();

assert_eq!(my_scope.get_value::<i64>("x").unwrap(), 123);

pub fn iter(&self) -> impl Iterator<Item = (&str, bool, Dynamic)>

Get an iterator to entries in the Scope. Shared values are flatten-cloned.

Example

use rhai::{Dynamic, Scope};

let mut my_scope = Scope::new();

my_scope.push("x", 42_i64);
my_scope.push_constant("foo", "hello");

let mut iter = my_scope.iter();

let (name, constant, value) = iter.next().unwrap();
assert_eq!(name, "x");
assert!(!constant);
assert_eq!(value.cast::<i64>(), 42);

let (name, constant, value) = iter.next().unwrap();
assert_eq!(name, "foo");
assert!(constant);
assert_eq!(value.cast::<String>(), "hello");

pub fn iter_raw<'x>(
&'x self
) -> impl Iterator<Item = (&'a str, bool, &'x Dynamic)> + 'x where
'x: 'a,

Get an iterator to entries in the Scope. Shared values are not expanded.