Struct casbin::rhai::TypeBuilder

pub struct TypeBuilder<'a, T>where
    T: Variant + Clone,{ /* private fields */ }

Builder to build the API of a custom type for use with an Engine.

The type is automatically registered when this builder is dropped.

Pretty-Print Name

By default the type is registered with Engine::register_type (i.e. without a pretty-print name).

To define a pretty-print name, call with_name, to use Engine::register_type_with_name instead.

Implementations

impl<'a, T> TypeBuilder<'a, T>where T: Variant + Clone,

pub fn with_name(&mut self, name: &'static str) -> &mut TypeBuilder<'a, T>

Set a pretty-print name for the type_of function.

pub fn on_print( &mut self, on_print: impl Fn(&mut T) -> String + SendSync + 'static ) -> &mut TypeBuilder<'a, T>

Pretty-print this custom type.

pub fn on_debug( &mut self, on_print: impl Fn(&mut T) -> String + SendSync + 'static ) -> &mut TypeBuilder<'a, T>

Debug-print this custom type.

pub fn with_fn<A, const N: usize, const C: bool, R, const L: bool>( &mut self, name: impl AsRef<str> + Into<SmartString<LazyCompact>>, method: impl RegisterNativeFunction<A, N, C, R, L> + SendSync + 'static ) -> &mut TypeBuilder<'a, T>where A: 'static, R: Variant + Clone,

Register a custom function.

impl<'a, T> TypeBuilder<'a, T>where T: Variant + Clone + IntoIterator, <T as IntoIterator>::Item: Variant + Clone,

pub fn is_iterable(&mut self) -> &mut TypeBuilder<'a, T>

Register a type iterator. This is an advanced API.

impl<'a, T> TypeBuilder<'a, T>where T: Variant + Clone,

pub fn with_get<const C: bool, V, const L: bool>( &mut self, name: impl AsRef<str>, get_fn: impl RegisterNativeFunction<(Mut<T>,), 1, C, V, L> + SendSync + 'static ) -> &mut TypeBuilder<'a, T>where V: Variant + Clone,

Register a getter function.

The function signature must start with &mut self and not &self.

Not available under no_object.

pub fn with_set<const C: bool, V, const L: bool>( &mut self, name: impl AsRef<str>, set_fn: impl RegisterNativeFunction<(Mut<T>, V), 2, C, (), L> + SendSync + 'static ) -> &mut TypeBuilder<'a, T>where V: Variant + Clone,

Register a setter function.

Not available under no_object.

pub fn with_get_set<const C1: bool, const C2: bool, V, const L1: bool, const L2: bool>( &mut self, name: impl AsRef<str>, get_fn: impl RegisterNativeFunction<(Mut<T>,), 1, C1, V, L1> + SendSync + 'static, set_fn: impl RegisterNativeFunction<(Mut<T>, V), 2, C2, (), L2> + SendSync + 'static ) -> &mut TypeBuilder<'a, T>where V: Variant + Clone,

Short-hand for registering both getter and setter functions.

All function signatures must start with &mut self and not &self.

Not available under no_object.

impl<'a, T> TypeBuilder<'a, T>where T: Variant + Clone,

pub fn with_indexer_get<X, const C: bool, V, const L: bool>( &mut self, get_fn: impl RegisterNativeFunction<(Mut<T>, X), 2, C, V, L> + SendSync + 'static ) -> &mut TypeBuilder<'a, T>where X: Variant + Clone, V: Variant + Clone,

Register an index getter.

The function signature must start with &mut self and not &self.

Not available under both no_index and no_object.

pub fn with_indexer_set<X, const C: bool, V, const L: bool>( &mut self, set_fn: impl RegisterNativeFunction<(Mut<T>, X, V), 3, C, (), L> + SendSync + 'static ) -> &mut TypeBuilder<'a, T>where X: Variant + Clone, V: Variant + Clone,

Register an index setter.

Not available under both no_index and no_object.

pub fn with_indexer_get_set<X, const C1: bool, const C2: bool, V, const L1: bool, const L2: bool>( &mut self, get_fn: impl RegisterNativeFunction<(Mut<T>, X), 2, C1, V, L1> + SendSync + 'static, set_fn: impl RegisterNativeFunction<(Mut<T>, X, V), 3, C2, (), L2> + SendSync + 'static ) -> &mut TypeBuilder<'a, T>where X: Variant + Clone, V: Variant + Clone,

Short-hand for registering both index getter and setter functions.

Not available under both no_index and no_object.

impl<'a, T> TypeBuilder<'a, T>where T: Variant + Clone,

pub fn with_result_fn<S, A, const N: usize, const C: bool, R, F>( &mut self, name: S, method: F ) -> &mut TypeBuilder<'a, T>where A: 'static, S: AsRef<str> + Into<SmartString<LazyCompact>>, R: Variant + Clone, F: RegisterNativeFunction<A, N, C, R, true> + SendSync + 'static,

👎Deprecated since 1.9.1: use with_fn instead

Register a custom fallible function.

Deprecated

This method is deprecated. Use with_fn instead.

This method will be removed in the next major version.

pub fn with_get_result<V, const C: bool>( &mut self, name: impl AsRef<str>, get_fn: impl RegisterNativeFunction<(Mut<T>,), 1, C, V, true> + SendSync + 'static ) -> &mut TypeBuilder<'a, T>where V: Variant + Clone,

👎Deprecated since 1.9.1: use with_get instead

Register a fallible getter function.

The function signature must start with &mut self and not &self.

Not available under no_object.

Deprecated

This method is deprecated. Use with_get instead.

This method will be removed in the next major version.

pub fn with_set_result<V, const C: bool>( &mut self, name: impl AsRef<str>, set_fn: impl RegisterNativeFunction<(Mut<T>, V), 2, C, (), true> + SendSync + 'static ) -> &mut TypeBuilder<'a, T>where V: Variant + Clone,

👎Deprecated since 1.9.1: use with_set instead

Register a fallible setter function.

Not available under no_object.

Deprecated

This method is deprecated. Use with_set instead.

This method will be removed in the next major version.

pub fn with_indexer_get_result<X, V, const C: bool>( &mut self, get_fn: impl RegisterNativeFunction<(Mut<T>, X), 2, C, V, true> + SendSync + 'static ) -> &mut TypeBuilder<'a, T>where X: Variant + Clone, V: Variant + Clone,

👎Deprecated since 1.9.1: use with_indexer_get instead

Register an fallible index getter.

The function signature must start with &mut self and not &self.

Not available under both no_index and no_object.

Deprecated

This method is deprecated. Use with_indexer_get instead.

This method will be removed in the next major version.

pub fn with_indexer_set_result<X, V, const C: bool>( &mut self, set_fn: impl RegisterNativeFunction<(Mut<T>, X, V), 3, C, (), true> + SendSync + 'static ) -> &mut TypeBuilder<'a, T>where X: Variant + Clone, V: Variant + Clone,

👎Deprecated since 1.9.1: use with_indexer_set instead

Register an fallible index setter.

Not available under both no_index and no_object.

Deprecated

This method is deprecated. Use with_indexer_set instead.

This method will be removed in the next major version.